requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
pydantic>=2.9.2
pydantic-settings>=2.5.2
sqlalchemy>=2.0.28
python-dotenv>=1.0.1
scikit-learn>=1.5.2
pandas>=2.2.3
networkx>=3.2.1
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
Pillow>=10.4.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.4
joblib>=1.4.2
roboflow>=1.1.50

